--- parsl/tests/test_aalst_patterns/__init__.py
@@ -1,4 +1,0 @@
-import os
-
-if str(os.environ.get('PARSL_TESTING', False)).lower() != 'true':
-    raise RuntimeError("must first run 'export PARSL_TESTING=True'")
